Refactor
--PostReducer
Changed some error on changing the state directly without the
reducer, specifically on the postList state.

--App.js
Changed the post route to match the Udacity requirements

--EditPost
Using the post_id param to check if it is a post to edit
or a createPost view.

--PostDetails
Sent deleteCOmmentById and setCurrentPost on created a Link
tag to navigate to EditPost when the user clicks on edit
button.